feat: Replace tsc/vite build split with vite-plugin-electron dev mode and HMR (no-changelog)
