+++ packages/@n8n/local-gateway/package.json
@@ -5,12 +5,11 @@
   "private": true,
   "main": "dist/main/index.js",
   "scripts": {
-    "build": "tsc -p tsconfig.build.json && vite build --config vite.renderer.config.mts && pnpm copy:assets",
-    "build:unchecked": "tsc -p tsconfig.build.json --noCheck && vite build --config vite.renderer.config.mts && pnpm copy:assets",
+    "build": "rimraf dist && tsc -p tsconfig.build.json --noEmit && vite build",
+    "build:unchecked": "rimraf dist && vite build",
     "lint": "eslint . --quiet",
     "typecheck": "tsc -p tsconfig.build.json --noEmit",
-    "copy:assets": "node scripts/copy-assets.js",
-    "dev": "pnpm copy:assets && concurrently \"vite build --watch --config vite.renderer.config.mts\" \"tsc -p tsconfig.build.json --watch\"",
+    "dev": "vite",
     "start": "electron dist/main/index.js",
     "dist:mac": "pnpm build && electron-builder --mac --config electron-builder.config.js",
     "dist:win": "pnpm build && electron-builder --win --config electron-builder.config.js",
@@ -38,13 +37,13 @@
     "@types/node": "catalog:",
     "@vitejs/plugin-vue": "catalog:frontend",
     "@vitest/coverage-v8": "catalog:",
-    "concurrently": "^8.2.0",
     "electron": "^41.3.0",
     "electron-builder": "^26.9.0",
     "rimraf": "catalog:",
     "sass-embedded": "catalog:",
     "unplugin-icons": "catalog:frontend",
     "vite": "catalog:",
+    "vite-plugin-electron": "catalog:",
     "vite-svg-loader": "catalog:frontend",
     "vitest": "catalog:"
   },
