--- packages/@n8n/local-gateway/scripts/copy-assets.js
@@ -1,8 +1,0 @@
-const fs = require('fs');
-
-function main() {
-	fs.mkdirSync('./dist/main', { recursive: true });
-	fs.cpSync('./assets', './dist/main/assets', { recursive: true });
-}
-
-main();
